register/pyproject.toml
[project]
name = "tavily-register"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyaml>=25.7.0",
    "pysocks>=1.7.1",
    "requests>=2.32.5",
    "reportlab>=4.4.9",
    "svglib>=1.6.0",
    "ddddocr<1.6.0",
    "numpy<2.0.0",
    "onnxruntime>=1.17.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
train = [
    "torch>=2.0",
    "torchvision>=0.15",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]
torchvision = [{ index = "pytorch-cu124" }]
